Tareas/Actividad8/requirements.txt
Flask==3.0.0
python-dotenv
Flask-Mail
celery
redis